Delivery Management Module Part 1
Added DeliveryManagementController

configured web.php to map the said controller of all its methods.

added folder in views for delivery management.

REMOVED MEAL-PROPOSAL-BLADE AND ORDERBASE-BLADE as they are not used for the project.

ADDED ORDERS TABLE AND MODEL
NOTE: no migration needed as they do not contain any thing as of now.